refactor browser components
